Add token list command
